// src/features/draft/commands/types/api-update-rows.handler.types.ts
import { GetRowByIdQueryReturnType } from 'src/features/row/queries/impl';
import { GetTableByIdReturnType } from 'src/features/table/queries/types';

export type ApiUpdateRowsHandlerReturnType = {
  table: NonNullable<GetTableByIdReturnType>;
  previousVersionTableId: string;
  rows: NonNullable<GetRowByIdQueryReturnType>[];
};